# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "rAsPI documentation",
    "version": "1.0.0",
    "description": "rAsPI is meant to be a general purpose API that can be easily modified and added to for use with raspberry pi projects."
  },
  "basePath": "/api",
  "host": "localhost:3000",
  "schemes": [
    "http"
  ],
  "paths": {
    "/answer/{id}": {
      "get": {
        "description": "Get answer by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id to be fetched.",
            "required": true,
            "type": "numeric"
          }
        ],
        "responses": {
          "200": {
            "description": "A JSON object representing an answer",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "An error string",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/example": {
      "get": {
        "description": "This is an example of how to execute a shell command that requires user input. The rAsPI project contains a directory called shell_scripts with a bash scripted called type_something.sh. When executed, it asks the user to type something and waits for input. Once the user types something and preses enter, it prints back what was typed. This endpoing shows how such a script can be executed via the api.",
        "responses": {
          "200": {
            "description": "A JSON object with property a property called result which contains an array of returned lines.",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "An error string",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/system": {
      "get": {
        "description": "This endpoint executes an awk script in the shell_scripts directory which parses a file to return the rAsPI version number.",
        "responses": {
          "200": {
            "description": "JSON object containing rAsPI version",
            "schema": {
              "type": "object",
              "properties": {
                "rAsPI_version": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "An error string",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "This endpoint allows you to post an allowed command (either sysShutdown or reboot) to be executed by the pi.",
        "parameters": [
          {
            "name": "command",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "type": "string",
                "description": "Command to execute on the system.",
                "required": true
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A string indicating the system status.",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "An error string",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}